LPS-196687 Move modal to fragments_and_widgets plugin
