+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/components/App.scss
@@ -12,7 +12,6 @@
 @import 'LayoutBreadcrumbs';
 @import 'LayoutViewport';
 @import 'NetworkstatusBar';
-@import 'ReorderSetModal';
 @import 'SaveFragmentCompositionModal';
 @import 'ShortcutModal';
 @import 'Sidebar';
