+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/plugins/fragments_and_widgets/components/_Plugin.scss
@@ -2,3 +2,5 @@
 @import 'TabCollection';
 @import 'TabItem';
 @import 'TabsPanel';
+
+@import 'reorder_sets_modal/ReorderSetsModal';
